requirements.txt
absl-py==2.3.0
accelerate==1.8.0
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiolimiter==1.2.1
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.9.0
apted==1.0.3
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
boto3==1.34.102
botocore==1.34.162
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contextgem==0.8.2
contourpy==1.3.2
cryptography==45.0.4
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.6.0
datefinder==0.7.3
debugpy==1.8.14
decorator==5.2.1
deepdoctection==0.43.2
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
Distance==0.1.3
distro==1.9.0
doclayout_yolo==0.0.2b1
docling==2.38.1
docling-core==2.38.2
docling-ibm-models==3.6.0
docling-parse==4.1.0
doctr==1.9.0
easydict==1.13
easyocr==1.7.2
effdet==0.4.1
einops==0.8.1
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.8.0.post1
fast-langdetect==0.2.5
fastapi==0.115.13
fasttext-predict==0.9.2.4
fasttext-wheel==0.9.2
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.1.1
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.58.4
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.3.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==1.34.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
GPUtil==1.4.0
gradio==5.34.2
gradio_client==1.10.3
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.0
grpcio-status==1.49.0rc1
h11==0.16.0
h5py==3.14.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-aiohttp==0.1.6
httpx-sse==0.4.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imagesize==1.4.1
img2table==1.0.5
importlib-metadata==6.11.0
importlib_resources==6.5.2
inflect==7.5.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jdeskew==0.3.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jpype1==1.5.2
json_repair==0.47.3
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.25
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.2
language_data==1.3.0
latex2mathml==3.78.0
layoutparser==0.3.4
lazy-imports==0.3.1
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
litellm==1.71.1
llvmlite==0.44.0
loguru==0.7.3
looseversion==1.3.0
lxml==5.4.0
magic-pdf==1.3.12
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
-e git+https://github.com/opendatalab/MinerU.git@ee93a714ac672c6416910fe2a247c3b5c523f51d#egg=mineru
ml_dtypes==0.5.1
mock==4.0.3
modelscope==1.27.1
more-itertools==10.7.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.12.2
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.43.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.0
openai==1.91.0
opencv-contrib-python==4.12.0.88
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.16.0
orjson==3.10.18
packaging==23.2
paddle2onnx==1.3.1
paddleclas==2.6.0
paddlefsl==1.1.0
paddlehub==2.4.0
paddlenlp==2.6.1
paddleocr==3.0.3
paddlepaddle==3.0.0
paddlex==3.0.3
pandas==2.3.0
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pdftext==0.6.3
pi_heif==0.22.0
pikepdf==9.9.0
Pillow==9.5.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
polars==1.31.0
poppler-utils==0.1.0
portalocker==3.2.0
premailer==3.10.0
preshed==3.0.10
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
prov==2.0.2
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.29
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
Pympler==1.1
PyMuPDF==1.24.14
pypandoc==1.15
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.9.0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-Levenshtein==0.27.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-ulid==3.0.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.0
rapid-table==2.0.2
RapidFuzz==3.13.0
rapidocr==3.2.0
rapidocr-onnxruntime==1.4.4
rarfile==4.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
robust-downloader==0.0.2
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.0
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
semchunk==2.2.2
sentencepiece==0.2.0
seqeval==1.2.2
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.4.9
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.25.0
streamlit-drawable-canvas==0.9.3
stringzilla==3.12.5
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
thinc==8.3.6
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.53.3
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==4.3.1
ujson==5.10.0
ultralytics==8.3.159
ultralytics-thop==2.0.14
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
visualdl==2.5.3
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
wtpsplit-lite==0.2.0
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
